+++ packages/common/src/objects/deep-set.ts
@@ -17,6 +17,7 @@ export function deepSet(
   let worker = target;
   const lastIdx = parts.length - 1;
   parts.forEach((p, idx) => {
+    // eslint-disable-next-line no-prototype-builtins
     if (!worker.hasOwnProperty(p) || worker[p] == null) {
       if (idx === lastIdx) {
         worker[p] = value;