+++ packages/sites/CHANGELOG.md
@@ -1,3 +1,22 @@
+# @esri/hub-sites [21.0.0](https://github.com/Esri/hub.js/compare/@esri/hub-sites@20.0.1...@esri/hub-sites@21.0.0) (2025-08-25)
+
+
+
+
+
+### Dependencies
+
+* **@esri/hub-common:** upgraded to 20.0.0
+* **@esri/hub-initiatives:** upgraded to 20.0.0
+* **@esri/hub-teams:** upgraded to 20.0.0
+
+## @esri/hub-sites [20.0.1](https://github.com/Esri/hub.js/compare/@esri/hub-sites@20.0.0...@esri/hub-sites@20.0.1) (2025-08-20)
+
+
+### Bug Fixes
+
+* add deprecation message to package README files ([#1972](https://github.com/Esri/hub.js/issues/1972)) ([98b8440](https://github.com/Esri/hub.js/commit/98b8440da176c5534e665f4b99c930cfc4115e52))
+
 # @esri/hub-sites [20.0.0](https://github.com/Esri/hub.js/compare/@esri/hub-sites@19.0.2...@esri/hub-sites@20.0.0) (2025-08-04)